fix(launch-modal): lift Identity RPC to layer + drop cancel-then-close double-dispatch
Two reagent findings on PR #910 round 2:

- [P1] TabModalLayer.tsx:268 — onCancel ran `req.onCancel()` (which
  `tabModal.replace`s back to Launch) AND `api.close()` synchronously,
  so close nullified the just-issued replace. Cancelling exited the
  launch flow instead of returning. Dropped `api.close()`; the caller
  is responsible for routing (replace vs close).

- [P1, from #911 review] AgentNewIdentityModal's panel-local
  `submitting()` wasn't visible to the layer's safeClose, so ESC /
  backdrop click mid-RPC unmounted the panel and the resolved promise
  then fired `onCreated`, re-opening Launch unexpectedly. Lifted the
  UpsertIdentityBundle RPC into TabModalLayer (mirrors the launch-
  agent dispatch pattern) so layer `submitting()` tracks the in-flight
  call. Also removed the panel's input-level Escape handler — the
  overlay's `safeClose` is now the single ESC path and respects the
  layer flag.

NewIdentityBundleRequest doc-comments updated to call out the
"caller owns the routing after onCreated/onCancel" contract so
future variants don't re-introduce the double-dispatch trap.
