+++ frontend/app/tab/tab-modal.ts
@@ -118,6 +118,12 @@ export interface AgentPrereqRequest {
  * "+ New" affordance on the Launch modal's Identity row creates an
  * empty Identity bundle. Connector setup (Claude/Codex/GitHub/AWS)
  * happens in the Identity pane afterward.
+ *
+ * The actual UpsertIdentityBundle RPC is owned by the layer so its
+ * `submitting()` flag (which gates safeClose) tracks the in-flight
+ * call — see reagent P1 on PR #911. Callers only supply the chain
+ * callbacks for after-success / on-cancel.
+ *
  * Phase β of SPEC_LAUNCH_MODAL_PROFILE_SECTION_2026_05_18.md.
  */
 export interface NewIdentityBundleRequest {
@@ -125,9 +131,15 @@ export interface NewIdentityBundleRequest {
     originBlockId: string;
     /** Initial value for the name field. Usually empty. */
     initialName?: string;
-    /** Called after the bundle is persisted on disk. The Launch
-     *  modal uses the id to auto-select on its next render. */
+    /** Called after the bundle is persisted on disk. Caller should
+     *  `tabModal.replace(launchRequest)` with the new id preselected;
+     *  the layer does NOT close after this fires. */
     onCreated: (bundleId: string, bundleName: string) => void;
+    /** Called when the user clicks Cancel. Caller should
+     *  `tabModal.replace(launchRequest)` with the prior selection
+     *  intact, OR `tabModal.close()` to exit. The layer does NOT
+     *  close after this fires — running both replace + close
+     *  synchronously nullified the replace, reagent P1 on PR #910. */
     onCancel: () => void;
 }
